Make brackets prettier and fix bug
Make brackets change their curvature based on the height of the
expression instead of having the same ends no matter how tall it is.

Fix a bug where the bracket height would not update when a nested
expression's height updates. This involves moving the call to
recomputeHeights() inside computeDimensions(), which could have
unintended consequences but nothing is crashing yet so I guess it's ok?